requirements.txt
Flask>=3.0.3
Flask_Limiter>=3.5.0
python-dotenv>=1.0.1
PyYAML>=6.0
Requests>=2.32.3